refactor: refactor validation, normalization and config clusters
multiple changes in here:
- the 'cluster' data normalization has been moved to the repository layer,
so all clusters have trailing slashes in their URLs when they are created or
updated in the DB, and a trailing slash is also added to the URL when a search
is performed
- the service layer "only" deals with validation of parameters, but allows empty
URLs which would be normalized at the repository layer (see above)

also, while refactoring some code, the old `configuration.ClusterConfig` type
was removed, so that from now on, the configuration returns a slice of
`repository.Cluster`, so there's no need to convert from one type to another.
Since we have some default values to apply (eg: `OSO` type of cluster and SA token
encrypted to `true`), a bit of boilerplate code was added to check the `optional`
and `default` tags on the `repository.Cluster` type to apply the default values
after the unmarshalling, using a custom `mapstructure` decoder.

Fixes #66

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>